requirements.txt
bar_chart_race==0.1.0
GitPython==3.1.43
GitPython==3.1.43
pandas==2.2.2
plotly==5.22.0
tqdm==4.64.1
tqdm==4.66.4
